Don't set `Content-Length` header for HTTP 204
According to RFC7230 (http://tools.ietf.org/html/rfc7230#section-3.3.1):

```
A server MUST NOT send a Content-Length header field in any response
   with a status code of 1xx (Informational) or 204 (No Content).  A
   server MUST NOT send a Content-Length header field in any 2xx
   (Successful) response to a CONNECT request (Section 4.3.6 of
   [RFC7231]).
```

Hobbit is automatically generating the `Content-Length` header for all
responses in `#finish`, which results in invalid headers in the case of
empty responses. These will trigger `Rack::Lint` errors as well.

This modifies the definition of `Hobbit::Response#finish` to skip setting
the `Content-Length` header in the case of 1xx or 204 response codes.